refactor(events): relocate mixin knowledge reactions to knowledge domain home (P4a Batch E2c)
Relocates the knowledge-coupled reactions embedded in the MIXED-domain
cross_subscribers handlers (basic/culture/strategic) to
aragora.knowledge.event_subscribers.KnowledgeEventSubscriber, completing
the events->knowledge edge clearance started by E2a/E2b (dedicated
knowledge_mound + validation modules):

- mound_to_culture, get_debate_culture_hints/_store_debate_culture
  (from CultureHandlersMixin)
- debate_outcome_to_knowledge, workflow_complete_to_supermemory,
  workflow_failed_to_supermemory, tier_demotion_to_revalidation,
  tier_promotion_to_knowledge (from BasicHandlersMixin)
- approval_to_km_reinforcement (from StrategicHandlersMixin)

Each mixin's other-domain reactions (E3 memory/reasoning/ranking, E4
debate, E5 workflow, E6 server) are left in place. manager.py registers
the relocated handlers via apply_registered_subscribers (E1 API);
get_cross_subscriber_manager() import path is unchanged.

Deletes the dead-duplicate events/subscribers/{debate,execution,
testfixer}_handlers.py (grok residual #2: zero live consumers, not
composed into CrossSubscriberManager) since they carry their own
aragora.knowledge/nomic/ranking imports blocking full edge clearance.
RELOCATE-UP NO-SHIM exemption applies (AGENTS.md / doc Sec8): no
re-export shim, no library/shims.md entry; all runtime + test consumers
(incl. @patch string paths) repointed to the new home.

Hand-shrinks the aragora.events -> aragora.knowledge baseline string in
scripts/baselines/import_contracts_baseline.json now that
check_import_contracts.py confirms zero remaining contributors
(frozen_from_ref unchanged). events->nomic and events->ranking also
cleared as a side effect of the dead-file deletion but their baseline
strings are left for a follow-up batch to keep this diff scoped to the
knowledge edge.
