+++ aragora/events/cross_subscribers/handlers/basic.py
@@ -515,57 +515,6 @@ class BasicHandlersMixin:
             f"consensus={consensus} confidence={confidence:.2f}"
         )
 
-    def _handle_debate_outcome_to_knowledge(self, event: StreamEvent) -> None:
-        """Debate end → Knowledge Mound outcome persistence.
-
-        When a debate ends, persist the outcome (winning position,
-        key arguments, consensus strength) into the Knowledge Mound
-        for future debate context enrichment.
-        """
-        data = event.data
-        debate_id = data.get("debate_id", "")
-        consensus = data.get("consensus_reached", False)
-        confidence = data.get("confidence", 0.0)
-        task = data.get("task", "")
-
-        if not consensus or confidence < 0.6:
-            return  # Only persist high-confidence outcomes
-
-        try:
-            from aragora.knowledge.mound import get_knowledge_mound
-
-            mound = get_knowledge_mound()
-            if mound is None:
-                return
-
-            outcome_content = {
-                "debate_id": debate_id,
-                "task": task[:500] if task else "",
-                "consensus_reached": consensus,
-                "confidence": confidence,
-                "winning_position": data.get("winning_position", ""),
-                "key_arguments": data.get("key_arguments", [])[:10],
-            }
-
-            import asyncio
-
-            item = {
-                "content": str(outcome_content),
-                "source": f"debate:{debate_id}",
-                "node_type": "debate_outcome",
-                "metadata": outcome_content,
-            }
-            try:
-                loop = asyncio.get_running_loop()
-                loop.create_task(mound.ingest(item))
-            except RuntimeError:
-                asyncio.run(mound.ingest(item))
-            logger.debug("Persisted debate outcome to KM: %s", debate_id)
-        except ImportError:
-            pass  # Knowledge Mound not available
-        except (RuntimeError, TypeError, AttributeError, ValueError, OSError) as e:
-            logger.debug("KM outcome persistence failed: %s", e)
-
     def _handle_debate_end_to_workflow(self, event: StreamEvent) -> None:
         """Debate end -> post-debate workflow automation.
 
@@ -590,158 +539,3 @@ class BasicHandlersMixin:
             logger.debug("PostDebateWorkflowSubscriber not available")
         except (KeyError, TypeError, AttributeError, ValueError) as e:
             logger.debug("Debate end -> workflow handler error: %s", e)
-
-    def _handle_workflow_outcome_to_supermemory(self, event: StreamEvent) -> None:
-        """Workflow completion/failure → Supermemory persistence.
-
-        When a workflow completes or fails, store the outcome in supermemory
-        for cross-workflow learning. This creates institutional memory of
-        what worked and what didn't, enabling future workflows to benefit
-        from past experience.
-        """
-        data = event.data
-        workflow_id = data.get("workflow_id", "")
-        definition_id = data.get("definition_id", "")
-        success = data.get("success", False)
-
-        if not workflow_id:
-            return
-
-        logger.info(
-            "Storing workflow outcome in supermemory: workflow=%s success=%s",
-            workflow_id,
-            success,
-        )
-
-        try:
-            from aragora.knowledge.mound import get_knowledge_mound
-
-            mound = get_knowledge_mound()
-            if mound is None:
-                return
-
-            outcome = {
-                "workflow_id": workflow_id,
-                "definition_id": definition_id,
-                "success": success,
-                "duration_ms": data.get("duration_ms", 0),
-                "steps_executed": data.get("steps_executed", 0),
-                "error": data.get("error", ""),
-            }
-
-            status = "completed successfully" if success else "failed"
-            content = (
-                f"Workflow {definition_id or workflow_id} {status}. "
-                f"Steps: {outcome['steps_executed']}, "
-                f"Duration: {outcome['duration_ms']}ms"
-            )
-            if not success and outcome["error"]:
-                content += f". Error: {outcome['error'][:200]}"
-
-            import asyncio
-
-            wf_item = {
-                "content": content,
-                "source": f"workflow:{workflow_id}",
-                "node_type": "workflow_outcome",
-                "metadata": outcome,
-            }
-            try:
-                loop = asyncio.get_running_loop()
-                loop.create_task(mound.ingest(wf_item))
-            except RuntimeError:
-                asyncio.run(mound.ingest(wf_item))
-            logger.debug("Workflow outcome stored in KM: %s", workflow_id)
-        except ImportError:
-            pass  # Knowledge Mound not available
-        except (RuntimeError, TypeError, AttributeError, ValueError, OSError) as e:
-            logger.debug("KM workflow storage failed: %s", e)
-
-    def _handle_tier_demotion_to_revalidation(self, event: StreamEvent) -> None:
-        """Memory tier demotion → Re-validation trigger.
-
-        When a memory entry is demoted to slow or glacial tier, trigger
-        re-validation to ensure the content is still accurate before
-        it becomes harder to access. This prevents stale or incorrect
-        knowledge from persisting in lower tiers without review.
-        """
-        data = event.data
-        memory_id = data.get("memory_id", "")
-        to_tier = data.get("to_tier", "")
-        from_tier = data.get("from_tier", "")
-
-        if not memory_id:
-            return
-
-        # Only re-validate on demotion to slow or glacial tiers
-        if to_tier not in ("slow", "glacial"):
-            return
-
-        logger.info(
-            "Tier demotion re-validation: memory=%s from=%s to=%s",
-            memory_id,
-            from_tier,
-            to_tier,
-        )
-
-        try:
-            from aragora.knowledge.mound import get_knowledge_mound
-
-            mound = get_knowledge_mound()
-            if mound is None:
-                return
-
-            # Mark for re-validation in KM
-            if hasattr(mound, "mark_for_revalidation"):
-                mound.mark_for_revalidation(
-                    source=f"continuum:{memory_id}",
-                    reason=f"tier_demotion:{from_tier}->{to_tier}",
-                )
-                logger.debug(
-                    "Marked memory %s for KM re-validation after demotion",
-                    memory_id,
-                )
-        except ImportError:
-            pass  # Knowledge Mound not available
-        except (RuntimeError, TypeError, AttributeError, ValueError) as e:
-            logger.debug("KM re-validation trigger failed: %s", e)
-
-    def _handle_tier_promotion_to_knowledge(self, event: StreamEvent) -> None:
-        """Memory tier promotion → Knowledge Mound notification.
-
-        When a memory entry is promoted to a faster tier, notify KM
-        so it can prioritize that knowledge for retrieval and ensure
-        the entry's importance is reflected in search rankings.
-        """
-        data = event.data
-        memory_id = data.get("memory_id", "")
-        to_tier = data.get("to_tier", "")
-        surprise_score = data.get("surprise_score", 0.0)
-
-        if not memory_id:
-            return
-
-        logger.debug(
-            "Tier promotion notification: memory=%s to=%s surprise=%.3f",
-            memory_id,
-            to_tier,
-            surprise_score,
-        )
-
-        try:
-            from aragora.knowledge.mound import get_knowledge_mound
-
-            mound = get_knowledge_mound()
-            if mound is None:
-                return
-
-            # Boost importance in KM based on promotion
-            if hasattr(mound, "boost_importance"):
-                mound.boost_importance(
-                    source=f"continuum:{memory_id}",
-                    factor=1.0 + surprise_score,
-                )
-        except ImportError:
-            pass  # Knowledge Mound not available
-        except (RuntimeError, TypeError, AttributeError, ValueError) as e:
-            logger.debug("KM importance boost failed: %s", e)
