--- aragora/events/subscribers/debate_handlers.py
@@ -1,509 +1,0 @@
-"""
-Debate and ELO event handler mixin for CrossSubscriberManager.
-
-Handles event flow related to debates:
-- ELO → Debate: Performance updates affect team selection
-- Calibration → Agent: Confidence adjustments
-- Culture → Debate: Organizational patterns
-- Staleness → Debate: Knowledge freshness checks
-"""
-
-from __future__ import annotations
-
-import asyncio
-import logging
-import time
-from typing import TYPE_CHECKING, Any
-from collections.abc import Callable
-
-from aragora.events.types import StreamEvent
-
-if TYPE_CHECKING:
-    pass
-
-logger = logging.getLogger(__name__)
-
-
-class DebateHandlersMixin:
-    """Mixin providing debate and ELO event handlers.
-
-    This mixin requires the implementing class to provide:
-    - stats: dict - Handler statistics tracking
-    - retry_handler: Any - Retry logic handler
-    - circuit_breaker: Any - Circuit breaker for failure protection
-    - _culture_cache: dict - Cache for culture hints
-    - _culture_cache_ttl: float - TTL for culture cache entries
-    - _staleness_debounce: dict - Debounce tracking for staleness events
-    - _staleness_debounce_seconds: float - Debounce interval
-    - _is_km_handler_enabled(handler_name: str) -> bool - Feature flag check
-    """
-
-    # Type annotations for required attributes from the implementing class
-    stats: dict
-    retry_handler: Any
-    circuit_breaker: Any
-    _culture_cache: dict
-    _culture_cache_ttl: float
-    _staleness_debounce: dict
-    _staleness_debounce_seconds: float
-
-    # Required method from parent class - checks feature flags
-    _is_km_handler_enabled: Callable[[str], bool]
-
-    def _handle_elo_to_debate(self, event: StreamEvent) -> None:
-        """Handle ELO → Debate events.
-
-        Updates team selection weights based on performance changes.
-        """
-        _check_and_record_slo: Callable[[str, float, str], Any] | None = None
-        try:
-            from aragora.observability.metrics.slo import check_and_record_slo as _slo_fn
-
-            _check_and_record_slo = _slo_fn
-        except ImportError:
-            pass
-
-        start = time.time()
-
-        try:
-            data = event.data
-            agent_name = data.get("agent_name")
-            new_rating = data.get("new_rating")
-            rating_change = data.get("rating_change", 0)
-            domain = data.get("domain", "general")
-
-            if not agent_name or new_rating is None:
-                logger.warning("ELO → Debate: Missing agent_name or new_rating")
-                return
-
-            # Update team selector weights
-            try:
-                from aragora.debate.team_selector import TeamSelector
-
-                selector = TeamSelector()
-                if hasattr(selector, "update_agent_weight"):
-                    selector.update_agent_weight(
-                        agent=agent_name,
-                        rating=new_rating,
-                        change=rating_change,
-                        domain=domain,
-                    )
-                    logger.debug(
-                        f"Updated team selection weight for {agent_name}: "
-                        f"{new_rating} ({'+' if rating_change >= 0 else ''}{rating_change})"
-                    )
-            except ImportError:
-                logger.debug("TeamSelector not available for ELO updates")
-            except (RuntimeError, TypeError, AttributeError, ValueError) as e:
-                logger.debug("TeamSelector update failed: %s", e)
-
-            self.stats["elo_to_debate"]["events"] += 1
-
-            # Record SLO
-            if _check_and_record_slo is not None:
-                latency_ms = (time.time() - start) * 1000
-                _check_and_record_slo("elo_to_debate", latency_ms, "success")
-
-        except (KeyError, TypeError, AttributeError, ValueError) as e:
-            logger.error("ELO → Debate handler error: %s", e)
-            self.stats["elo_to_debate"]["errors"] += 1
-
-    def _handle_calibration_to_agent(self, event: StreamEvent) -> None:
-        """Handle Calibration → Agent events.
-
-        Updates agent confidence based on calibration results.
-        """
-        _check_and_record_slo: Callable[[str, float, str], Any] | None = None
-        try:
-            from aragora.observability.metrics.slo import check_and_record_slo as _slo_fn
-
-            _check_and_record_slo = _slo_fn
-        except ImportError:
-            pass
-
-        start = time.time()
-
-        try:
-            data = event.data
-            agent_name = data.get("agent_name")
-            calibration_score = data.get("calibration_score")
-            overconfidence = data.get("overconfidence", 0)
-            sample_size = data.get("sample_size", 0)
-
-            if not agent_name or calibration_score is None:
-                return
-
-            # Update agent confidence scaling
-            try:
-                from aragora.debate.confidence import ConfidenceManager
-
-                manager = ConfidenceManager()
-                if hasattr(manager, "update_agent_calibration"):
-                    manager.update_agent_calibration(
-                        agent=agent_name,
-                        score=calibration_score,
-                        overconfidence=overconfidence,
-                        sample_size=sample_size,
-                    )
-                    logger.debug(
-                        f"Updated calibration for {agent_name}: "
-                        f"score={calibration_score:.2f}, overconfidence={overconfidence:.2f}"
-                    )
-            except ImportError:
-                logger.debug("ConfidenceManager not available")
-            except (RuntimeError, TypeError, AttributeError, ValueError) as e:
-                logger.debug("Calibration update failed: %s", e)
-
-            self.stats["calibration_to_agent"]["events"] += 1
-
-            # Record SLO
-            if _check_and_record_slo is not None:
-                latency_ms = (time.time() - start) * 1000
-                _check_and_record_slo("calibration_to_agent", latency_ms, "success")
-
-        except (KeyError, TypeError, AttributeError, ValueError) as e:
-            logger.error("Calibration → Agent handler error: %s", e)
-            self.stats["calibration_to_agent"]["errors"] += 1
-
-    def _handle_culture_to_debate(self, event: StreamEvent) -> None:
-        """Handle Culture → Debate events.
-
-        Applies organizational debate patterns.
-        """
-        if not self._is_km_handler_enabled("culture_to_debate"):
-            return
-
-        try:
-            data = event.data
-            debate_id = data.get("debate_id")
-            culture_hints = data.get("hints", {})
-            workspace = data.get("workspace", "default")
-
-            if not debate_id:
-                return
-
-            # Cache culture hints for the debate
-            cache_key = f"{workspace}:{debate_id}"
-            self._culture_cache[cache_key] = {
-                "hints": culture_hints,
-                "timestamp": time.time(),
-            }
-
-            logger.debug("Cached culture hints for debate %s", debate_id)
-            self.stats["culture_to_debate"]["events"] += 1
-
-        except (KeyError, TypeError, AttributeError, ValueError) as e:
-            logger.error("Culture → Debate handler error: %s", e)
-            self.stats["culture_to_debate"]["errors"] += 1
-
-    def _handle_mound_to_culture(self, event: StreamEvent) -> None:
-        """Handle Knowledge Mound → Culture events.
-
-        Retrieves organizational patterns for debate context.
-        """
-        if not self._is_km_handler_enabled("mound_to_culture"):
-            return
-
-        try:
-            data = event.data
-            debate_id = data.get("debate_id")
-            task = data.get("task")
-            workspace = data.get("workspace", "default")
-
-            if not task:
-                return
-
-            # Async retrieval of culture patterns
-            async def retrieve_culture():
-                try:
-                    from aragora.knowledge.mound_core import KnowledgeMound
-
-                    mound = KnowledgeMound(workspace_id=workspace)
-                    if hasattr(mound, "get_culture_patterns"):
-                        patterns = await mound.get_culture_patterns(task)
-                        if patterns and debate_id:
-                            cache_key = f"{workspace}:{debate_id}"
-                            self._culture_cache[cache_key] = {
-                                "hints": patterns,
-                                "timestamp": time.time(),
-                            }
-                except (
-                    ImportError,
-                    RuntimeError,
-                    TypeError,
-                    AttributeError,
-                    ValueError,
-                    OSError,
-                ) as e:
-                    logger.debug("Culture retrieval failed: %s", e)
-
-            try:
-                asyncio.get_running_loop()
-                asyncio.create_task(retrieve_culture())
-            except RuntimeError:
-                # No event loop, skip async operation
-                pass
-
-            self.stats["mound_to_culture"]["events"] += 1
-
-        except (KeyError, TypeError, AttributeError, ValueError, RuntimeError) as e:
-            logger.error("Mound → Culture handler error: %s", e)
-            self.stats["mound_to_culture"]["errors"] += 1
-
-    def _store_debate_culture(
-        self,
-        debate_id: str,
-        task: str,
-        patterns: dict,
-        workspace: str = "default",
-    ) -> None:
-        """Store debate culture patterns for later retrieval.
-
-        Args:
-            debate_id: Debate identifier
-            task: Debate task/topic
-            patterns: Culture patterns to store
-            workspace: Workspace identifier
-        """
-        cache_key = f"{workspace}:{debate_id}"
-        self._culture_cache[cache_key] = {
-            "hints": patterns,
-            "task": task,
-            "timestamp": time.time(),
-        }
-        logger.debug("Stored culture patterns for debate %s", debate_id)
-
-    def get_debate_culture_hints(self, debate_id: str, workspace: str = "default") -> dict:
-        """Get cached culture hints for a debate.
-
-        Args:
-            debate_id: Debate identifier
-            workspace: Workspace identifier
-
-        Returns:
-            Culture hints dict or empty dict if not found/expired
-        """
-        cache_key = f"{workspace}:{debate_id}"
-        entry = self._culture_cache.get(cache_key)
-
-        if not entry:
-            return {}
-
-        # Check TTL
-        if time.time() - entry["timestamp"] > self._culture_cache_ttl:
-            del self._culture_cache[cache_key]
-            return {}
-
-        return entry.get("hints", {})
-
-    def _handle_staleness_to_debate(self, event: StreamEvent) -> None:
-        """Handle Staleness → Debate events.
-
-        Triggers knowledge refresh when stale data detected.
-        """
-        if not self._is_km_handler_enabled("staleness_to_debate"):
-            return
-
-        try:
-            data = event.data
-            node_ids = data.get("stale_nodes", [])
-            workspace = data.get("workspace", "default")
-
-            if not node_ids:
-                return
-
-            # Debounce to avoid refresh storms
-            debounce_key = f"{workspace}:staleness"
-            last_check = self._staleness_debounce.get(debounce_key, 0)
-            now = time.time()
-
-            if now - last_check < self._staleness_debounce_seconds:
-                logger.debug("Staleness check debounced")
-                return
-
-            self._staleness_debounce[debounce_key] = now
-
-            # Trigger refresh for stale nodes
-            try:
-                from aragora.knowledge.mound_core import KnowledgeMound
-
-                mound = KnowledgeMound(workspace_id=workspace)
-                if hasattr(mound, "refresh_nodes"):
-                    mound.refresh_nodes(node_ids)
-                    logger.debug("Triggered refresh for %s stale nodes", len(node_ids))
-            except (ImportError, RuntimeError, TypeError, AttributeError, ValueError, OSError) as e:
-                logger.debug("Staleness refresh failed: %s", e)
-
-            self.stats["staleness_to_debate"]["events"] += 1
-
-        except (KeyError, TypeError, AttributeError, ValueError) as e:
-            logger.error("Staleness → Debate handler error: %s", e)
-            self.stats["staleness_to_debate"]["errors"] += 1
-
-    def _handle_provenance_to_mound(self, event: StreamEvent) -> None:
-        """Handle Provenance → Knowledge Mound events.
-
-        Syncs claim provenance data for cross-debate verification.
-        """
-        if not self._is_km_handler_enabled("provenance_to_mound"):
-            return
-
-        try:
-            data = event.data
-            claim_id = data.get("claim_id")
-            sources = data.get("sources", [])
-            verification_status = data.get("verification_status")
-            workspace = data.get("workspace", "default")
-
-            if not claim_id:
-                return
-
-            try:
-                from aragora.knowledge.mound_core import KnowledgeMound
-
-                mound = KnowledgeMound(workspace_id=workspace)
-                if hasattr(mound, "update_provenance"):
-                    mound.update_provenance(
-                        claim_id=claim_id,
-                        sources=sources,
-                        verification_status=verification_status,
-                    )
-                    logger.debug("Updated provenance for claim %s", claim_id)
-            except (ImportError, RuntimeError, TypeError, AttributeError, ValueError, OSError) as e:
-                logger.debug("Provenance update failed: %s", e)
-
-            self.stats["provenance_to_mound"]["events"] += 1
-
-        except (KeyError, TypeError, AttributeError, ValueError) as e:
-            logger.error("Provenance → Mound handler error: %s", e)
-            self.stats["provenance_to_mound"]["errors"] += 1
-
-    def _handle_debate_outcome_to_knowledge(self, event: StreamEvent) -> None:
-        """Handle DEBATE_END events → persist outcome to Knowledge Mound.
-
-        Extracts the winning position, key arguments, and consensus strength
-        from completed debates and stores them as knowledge facts. This enables
-        future debates on similar topics to retrieve prior conclusions.
-        """
-        if not self._is_km_handler_enabled("debate_outcome_to_knowledge"):
-            return
-
-        try:
-            data = event.data
-            debate_id = data.get("debate_id", "")
-            task = data.get("task", "")
-            consensus_reached = data.get("consensus_reached", False)
-            confidence = data.get("confidence", 0.0)
-            synthesis = data.get("synthesis", "")
-            winning_position = data.get("winning_position", "")
-            workspace = data.get("workspace", "default")
-
-            if not task or not debate_id:
-                return
-
-            # Only persist outcomes with meaningful consensus
-            if not consensus_reached and confidence < 0.5:
-                return
-
-            content = winning_position or synthesis
-            if not content:
-                return
-
-            # Store as knowledge fact
-            try:
-                from aragora.knowledge.mound_core import KnowledgeMound
-
-                mound = KnowledgeMound(workspace_id=workspace)
-                if hasattr(mound, "add_fact"):
-                    mound.add_fact(
-                        content=content[:2000],
-                        source_type="CONSENSUS",
-                        source_id=debate_id,
-                        domain=data.get("domain", "general"),
-                        confidence=confidence,
-                        metadata={
-                            "debate_id": debate_id,
-                            "task": task[:500],
-                            "consensus_reached": consensus_reached,
-                        },
-                    )
-                    logger.debug(
-                        "Stored debate outcome as knowledge: debate=%s confidence=%.2f",
-                        debate_id,
-                        confidence,
-                    )
-            except ImportError:
-                logger.debug("KnowledgeMound not available for debate outcome storage")
-            except (RuntimeError, TypeError, AttributeError, ValueError, OSError) as e:
-                logger.debug("Knowledge storage failed: %s", e)
-
-            self.stats.setdefault("debate_outcome_to_knowledge", {"events": 0, "errors": 0})
-            self.stats["debate_outcome_to_knowledge"]["events"] += 1
-
-        except (KeyError, TypeError, AttributeError, ValueError) as e:
-            logger.error("Debate outcome → Knowledge handler error: %s", e)
-            self.stats.setdefault("debate_outcome_to_knowledge", {"events": 0, "errors": 0})
-            self.stats["debate_outcome_to_knowledge"]["errors"] += 1
-
-    def _handle_debate_end_to_workflow(self, event: StreamEvent) -> None:
-        """Handle DEBATE_END events -> trigger post-debate workflow automation.
-
-        Delegates to PostDebateWorkflowSubscriber to classify the debate
-        outcome and trigger the appropriate workflow template (e.g.,
-        implementation PR, follow-up review, escalation, or retry).
-        """
-        try:
-            from aragora.events.subscribers.workflow_automation import (
-                PostDebateWorkflowSubscriber,
-            )
-
-            subscriber = PostDebateWorkflowSubscriber()
-            subscriber.handle_debate_end(event)
-
-            self.stats.setdefault("debate_end_to_workflow", {"events": 0, "errors": 0})
-            self.stats["debate_end_to_workflow"]["events"] += 1
-
-            # Propagate error count from the subscriber
-            if subscriber.stats["errors"] > 0:
-                self.stats["debate_end_to_workflow"]["errors"] += subscriber.stats["errors"]
-
-        except ImportError:
-            logger.debug("PostDebateWorkflowSubscriber not available")
-        except (KeyError, TypeError, AttributeError, ValueError) as e:
-            logger.error("Debate End -> Workflow handler error: %s", e)
-            self.stats.setdefault("debate_end_to_workflow", {"events": 0, "errors": 0})
-            self.stats["debate_end_to_workflow"]["errors"] += 1
-
-    def _handle_mound_to_provenance(self, event: StreamEvent) -> None:
-        """Handle Knowledge Mound → Provenance events.
-
-        Provides historical provenance data for claim verification.
-        """
-        if not self._is_km_handler_enabled("mound_to_provenance"):
-            return
-
-        try:
-            data = event.data
-            claim_text = data.get("claim_text")
-            debate_id = data.get("debate_id")
-            workspace = data.get("workspace", "default")
-
-            if not claim_text:
-                return
-
-            try:
-                from aragora.knowledge.mound_core import KnowledgeMound
-
-                mound = KnowledgeMound(workspace_id=workspace)
-                if hasattr(mound, "find_claim_provenance"):
-                    provenance = mound.find_claim_provenance(claim_text)
-                    if provenance:
-                        logger.debug("Found provenance for claim in debate %s", debate_id)
-            except (ImportError, RuntimeError, TypeError, AttributeError, ValueError, OSError) as e:
-                logger.debug("Provenance lookup failed: %s", e)
-
-            self.stats["mound_to_provenance"]["events"] += 1
-
-        except (KeyError, TypeError, AttributeError, ValueError) as e:
-            logger.error("Mound → Provenance handler error: %s", e)
-            self.stats["mound_to_provenance"]["errors"] += 1
